Split instance_clone resource.go into per-operation files
Per the provider file-structure convention (split CRUD once resource.go
exceeds 250 lines), move the Create, Read, Update and Delete operations
and their helpers into create.go, read.go, update.go and delete.go.
resource.go retains Metadata, Schema, the resource wiring and small
shared helpers. No behaviour change.
